+++ changelog.md
@@ -6,6 +6,7 @@
    + Note: RCLootCouncil cannot track if these items will be Warforged/Titanforged. Only the guaranteed minimum ilvl is used.
 * There's now one of each item type included in the test items.
 * The class icon can now be replaced with spec icon.
+* Filter buttons' text now change color to indicate a filter is active.
 
 
 * **Localization**
@@ -26,6 +27,7 @@
    + E.g. the council can now see a candidates gear and ilvl before a candidates responds.
 * RCLootCouncil now sends the gear a candidate had equipped during the most recent encounter instead of the gear equipped when rolling.
    + This way candidates can't change their gear to appear to have lower ilvl.
+* If enabled, relic responses can now be filtered separately.
 
 
 * **Rolls**